Landing pages. html
A landing page, in the context of web development, refers to a single web page that serves as the entry point for a website or a specific marketing campaign. Landing pages are designed with a focused objective, such as promoting a product, service, event, or capturing leads. They are crafted to encourage visitors to take a specific action, like making a purchase, signing up for a newsletter, or filling out a form.